services/saw_api/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
psycopg[binary]==3.2.13
pgvector==0.3.6
pydantic==2.10.4
PyYAML==6.0.2
openai==1.58.1


